% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionLink.R
\name{actionLink}
\alias{actionLink}
\title{Action Link}
\usage{
actionLink(inputId, label, icon = NULL, ...)
}
\arguments{
\item{inputId}{This is the id of the link that you are creating.}

\item{label}{This is the label of the link that is visible to the user.}

\item{icon}{This is optional and adds an icon to the link.}
}
\description{
This function creates a link that acts as an action button (useful for creating an action button in a dropdown list)
}
\examples{
  actionLink("sample-id","ID Form", icon("id-card-o"))
}
\keyword{actionButton,}
\keyword{dropdown}
\keyword{link,}
\keyword{menu}
\keyword{modal,}
