% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box}
\alias{box}
\title{Box}
\usage{
box(..., title = NULL, footer = NULL, status = NULL,
  solidHeader = FALSE, background = NULL, width = 12, height = NULL,
  collapsible = TRUE, collapsed = FALSE)
}
\arguments{
\item{...}{UI code that belongs inside of box.}

\item{title}{Title of the box}

\item{footer}{Text that should appear in the footer if desired}

\item{status}{Status of the box (warning, normal, etc.)}

\item{collapsible}{Whether the box is collapsible or not}

\item{collapsed}{Whether the box is collapsed on page load by default}
}
\description{
This function creates a box which is collapsible.
}
\examples{
box(title = "Test Title", collapsible = T, uiOutput("Some Output"))
}
\keyword{UI}
\keyword{box,}
