% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxUI.R
\name{boxUI}
\alias{boxUI}
\title{Box UI}
\usage{
boxUI(title, output)
}
\arguments{
\item{title}{Title of the box}

\item{output}{UI code that belongs inside of box.}
}
\description{
This function creates a box.
}
\examples{
boxUI(title = "Test Title", uiOutput("Some Output"))
}
\keyword{UI}
\keyword{box,}
