% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonModal.R
\name{buttonModal}
\alias{buttonModal}
\title{Button Modal}
\usage{
buttonModal(inputId, label)
}
\arguments{
\item{inputId}{This is the id of the button.}

\item{label}{This is the label of the button visible to the user.}
}
\description{
This function creates a button within a modal that simultaneously closes the current modal and then opens another modal.
}
\examples{
buttonModal("upload-ui", "Upload Data")
}
\keyword{buttons,}
\keyword{interface}
\keyword{modal,}
\keyword{user}
