% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonRow.R
\name{buttonRow}
\alias{buttonRow}
\title{Button Row}
\usage{
buttonRow(...)
}
\arguments{
\item{...}{Insert button functions in a tagList {shiny}}
}
\description{
This function creates a row of buttons
}
\examples{
  buttonRow(taglist(
  buttonModal(buttonModal("sample-id","ID Form")),
  actionButton("submit","Submit Responses", icon = icon("check"))
  ))
}
\keyword{button}
\keyword{buttons}
\keyword{row,}
