% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castDataFunction.R
\name{CastData}
\alias{CastData}
\title{Cast Data Function (CRUD)}
\usage{
CastData(data)
}
\arguments{
\item{data}{Data to be cast.}
}
\description{
This function casts data as part of a CRUD series of functions.
}
\examples{
CreateDefaultRecord <- function() {
mydefault <- CastData(list(id = "0",
                           propertyForm = "",
                           yearForm = strftime(Sys.Date(), "\%Y"),
                           monthForm = strftime(Sys.Date(), "\%B"),
                           categoryForm = "",
                           notes = "",
                           requesterName = userInfo()$Name))
return (mydefault)
}
}
\keyword{CRUD,}
\keyword{cast}
