% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeToSeries.R
\name{changeToSeries}
\alias{changeToSeries}
\title{Convert a Data Frame to a Series for Use in Highcharter}
\usage{
changeToSeries(df, keyColumnNameAsText)
}
\arguments{
\item{df}{This is the data source to be converted}
}
\description{
This function converts a data frame to a series for use in highcharter
}
\examples{
dataSeries <- changeToSeries(sampleDF,keyColumn)
}
\keyword{data}
\keyword{manipulation}
