% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAreaChart.R
\name{createAreaChart}
\alias{createAreaChart}
\title{Create an Area Chart in Highcharter}
\usage{
createAreaChart(title, qbSource, dat, nameLabel, valueLabel, categoryLabel)
}
\arguments{
\item{title}{This is the title of the chart}

\item{qbSource}{This is the source html link}

\item{dat}{This is the data source to be graphed}

\item{nameLabel}{This is the name of the column that contains the "names" of the data series}

\item{valueLabel}{This is the name of the column that contains the "data" of the data series}

\item{categoryLabel}{This is the name of the column that contains the "categories" or x axis of the data series}
}
\description{
This function creates an area chart in highcharter
}
\examples{
createAreaChart("Iris Data Set Area Chart Example","https://en.wikipedia.org/wiki/Iris_flower_data_set",iris,"Species","Sepal.Length","Petal.Length")
}
\keyword{conversion}
\keyword{data}
\keyword{datetime}
\keyword{manipulation,}
