% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createColumn.R
\name{createColumn}
\alias{createColumn}
\title{Create a Column Chart Using Highcharter}
\usage{
createColumn(seriesName, categories, title, subTitle)
}
\arguments{
\item{seriesName}{This is the series that will be the data basis of the graph.}

\item{categories}{These are the names for each category.}

\item{title}{This is the title of the chart}

\item{subTitle}{This is the subtitle of the chart}
}
\description{
This function creates a column chart using highcharter.
}
\examples{
dat2 <- mtcars \%>\% group_by(cyl,am) \%>\% summarise(AvgMPG = mean(mpg)) \%>\% as.data.frame() \%>\% rename(Value = AvgMPG) \%>\% mutate(am = ifelse(am == 1,"Automatic","Manual"))
datSeries <- changeToSeries(dat2,"am")
createColumn(datSeries,dat2$cyl,"Fuel Efficiency","Miles per Gallon by Transmission and Cylinders")
}
\keyword{chart}
\keyword{column}
\keyword{highcharter,}
