% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createData.R
\name{CreateData}
\alias{CreateData}
\title{Create Data Function (CRUD)}
\usage{
CreateData(data)
}
\arguments{
\item{data}{Data to be added/created.}
}
\description{
This function appends or creates data.
}
\examples{
CreateData <- function(data) {
data <- CastData(data)
rownames(data) <- GetNextId()
if (exists("responses")) {
  responses <<- rbind(responses, data)
} else {
  responses <<- data
}
}
}
\keyword{CRUD,}
\keyword{create}
