% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDrilldown.R
\name{createDrilldown}
\alias{createDrilldown}
\title{Create a Drilldown Chart Using Highcharter}
\usage{
createDrilldown(df, idColumnNameAsText, title, subTitle)
}
\arguments{
\item{df}{This is the data frame that provides data for the graph.}

\item{idColumnNameAsText}{This is the name of the column that should be the "id" part of the series}

\item{title}{This is the title of the chart}

\item{subTitle}{This is the subtitle of the chart}
}
\description{
This function creates a column chart using highcharter.
}
\examples{
df <- data.frame(
name = c("Toyota", "Ford", "Volkswagen", "Cats", "Dogs", "Cows", "Sheep", "Pigs", "Apple", "Oranges"),
value = c(4, 2, 2, 4, 3, 1, 2, 1, 4, 2),
id = c(rep("cars",3), rep("animals",5), rep("fruits",2)),
stringsAsFactors = FALSE
)
ds <- summarizeDrilldownSeries(df,idColumnNameAsText)
datSeries <- changeToDrilldownSeries(df,idColumnNameAsText)
hc <- createDrilldown(df,idColumnNameAsText,"Drilldown Test","Some additional text here for subtitle")
}
\keyword{chart}
\keyword{drilldown}
\keyword{highcharter,}
