% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHeatmap.R
\name{createHeatmap}
\alias{createHeatmap}
\title{Create a Heatmap Using Highcharter}
\usage{
createHeatmap(data, xText, yText, valueText, title, subTitle, xTitle, yTitle)
}
\arguments{
\item{data}{This is the data source for the heatmap.}

\item{xText}{This is the column name in the data of the x values in the heatmap.}

\item{yText}{This is the column name in the data of the y values in the heatmap.}

\item{valueText}{This is the column name in the data of the data values in the heatmap.}

\item{title}{This is the title of the heatmap.}

\item{subTitle}{This is the subtitle of the heatmap.}

\item{xTitle}{This is the title of the x axis.}

\item{yTitle}{This is the title of the y axis.}
}
\description{
This function creates a heatmap using highcharter.
}
\examples{
output$heatmap <- highcharter::renderHighchart({
hc <- createHeatmap(dat,"dayOfWeek","hourOfDay","Value",
  "Heatmap of Marginal Trips by Hour and Day",
  "Heatmap values represent the marginal trips from one additional driver hour or rider session","Day of Week","Hour of the Day")
})
}
\keyword{heatmap}
\keyword{highcharter,}
