% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchModal.R
\name{createLaunchModal}
\alias{createLaunchModal}
\title{Create Launch Modal}
\usage{
createLaunchModal(inputId, label)
}
\arguments{
\item{inputId}{This is the id of the modal that you are creating.}

\item{label}{The label of the modal that appears to the user}
}
\description{
This function creates a modal that appears on launch
}
\examples{
createLaunchModal("upload-ui","Step 2: Upload data")
}
\keyword{modal}
