% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchModalOnPageLoad.R
\name{createLoadModalScript}
\alias{createLoadModalScript}
\title{Launch a Modal on Shiny Page Load Function}
\usage{
createLoadModalScript(modalTag)
}
\arguments{
\item{modalTag}{This is the id of the modal that you want launched on page load.}
}
\description{
This function loads a modal after a page is loaded in Shiny.
}
\examples{
createLoadModalScript("testModalName")
}
\keyword{modal}
