% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrezModal.R
\name{createPrezModal}
\alias{createPrezModal}
\title{Launch a Modal on Shiny Page Load Function}
\usage{
createPrezModal(id, modalTitle, prezSrc)
}
\arguments{
\item{id}{This is the id of the modal that you are creating.}

\item{modalTitle}{This is the title of the modal that appears at the top of the modal.}

\item{prezSrc}{This is the URL of the published Google presentation that you want to appear in the modal.}
}
\description{
This function loads a modal after a page is loaded in Shiny.
}
\examples{
output$postmortemModal <- renderUI({
createPrezModal("sample","Sample Prezentation Modal",
                "https://docs.google.com/presentation/d/14H24z59wQVTnAg9AKD_bIZ6F3DKnGIy9haBK557XMVo/embed?start=false&loop=false&delayms=3000")
})
}
\keyword{Google}
\keyword{Sheets}
\keyword{modal,}
