% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQuery.R
\name{createQuery}
\alias{createQuery}
\title{Create an Atlantis, Warehouse, or Hive QueryRunner function}
\usage{
createQuery(dataSourceName)
}
\arguments{
\item{dataSourceName}{Name of data source:  warehouse, hive, atlantis}
}
\description{
Query a variety of different databases
}
\examples{
qf <- createQuery("warehouse")
qf("https://qb.uberinternal.com/querybuilder/builder/dFuvjmAdo5",city_id = c(40,40), vehicle_view = c(1427,2177,3213))
}
\keyword{sql}
