% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUIModal.R
\name{createUIModal}
\alias{createUIModal}
\title{Create a Modal Based on a User Interface}
\usage{
createUIModal(id, modalTitle, uiOutputText)
}
\arguments{
\item{id}{This is the id of the modal that you are creating.}

\item{modalTitle}{This is the title of the modal that appears at the top of the modal.}

\item{uiOutputText}{This can be a tag list or a uiOutput from some other Shiny output.}
}
\description{
This function creates a modal based on a Shiny UI.
}
\examples{
  createUIModal("sample-id","Upload file modal",
                tagList(
                  fileInput("file1", "Load a File:")
                ))
}
\keyword{UI}
\keyword{modal,}
