% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWaterfall.R
\name{createWaterfall}
\alias{createWaterfall}
\title{Create a Waterfall Chart Using Highcharter}
\usage{
createWaterfall(dfNames, dfValues, isSum, isIntermediateSum, title, subTitle,
  xTitle, yTitle)
}
\arguments{
\item{dfNames}{These are the names of the categories}

\item{dfValues}{These are the values for each category.  Intermediate and sum categories should be left as NA.}

\item{isSum}{This is a boolean list of whether the names and values represent a sum in the waterfall chart.}

\item{isIntermediateSum}{This is a boolean list of whether the names and values represent an intermediate sum in the waterfall chart.}

\item{title}{This is the title of the chart}

\item{subTitle}{This is the subtitle of the chart}

\item{xTitle}{This is the title of the x axis.}

\item{yTitle}{This is the title of the y axis.}
}
\description{
This function creates a waterfall chart using highcharter.  It automatically converts entered values into percentages (percentages are calculated relative to the first value entered).  It also automatically converts entered values into a series for plotting and also automatically chooses the color for a given category based on the category's value (positive or negative).
}
\examples{
dfNames <- c("Trips where passengers paid a surcharge",
"Double counted trips (pickup AND dropoff in special venue)",
"Trips where Uber paid a surcharge and didn't charge the rider",
"Trips where the dropoff did not occur in the current month",
"'Line' trips","Trips where Uber paid a $5 fee to the city")
dfValues <- c(416468,2338,140,72,-1773,NA)
isIntermediateSum = c(0,0,0,0,0,0)
isSum = c(0,0,0,0,0,1)
output$heatmap <- highcharter::renderHighchart({
createWaterfall(dfNames, dfValues, isSum, isIntermediateSum,"Line Trip Reconciliation",
"Cumulative Effect of Adjustments","Adjustment Type","Percent Effect")
})
}
\keyword{highcharter,}
\keyword{waterfall}
