% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvFile.R
\name{csvFile}
\alias{csvFile}
\title{CSV File Handler}
\usage{
csvFile(input, output, session, stringsAsFactors = FALSE)
}
\description{
This function handles the input of the CSV file input.  Needed for CSV File Input
}
\examples{
datafile <- callModule(csvFile, "datafile", stringsAsFactors = FALSE)
}
\keyword{csv,}
\keyword{file}
\keyword{input,}
