% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvFileInput.R
\name{csvFileInput}
\alias{csvFileInput}
\title{CSV File Input}
\usage{
csvFileInput(id, label = "CSV file")
}
\arguments{
\item{id}{This is the id of the input.}

\item{label}{This is the label of the input.}
}
\description{
This function creates a csv file input.
}
\examples{
csvFileInput("property","Upload CSV of Properties")
}
\keyword{csv,}
\keyword{file}
\keyword{input,}
