% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteData.R
\name{DeleteData}
\alias{DeleteData}
\title{Delete Data Function (CRUD)}
\usage{
DeleteData(data)
}
\arguments{
\item{data}{Data to be deleted}
}
\description{
This function deletes the record selected
}
\examples{
observeEvent(input$delete, {
  DeleteData(formData())
  UpdateInputs(CreateDefaultRecord(), session)
}, priority = 1)
}
\keyword{CRUD,}
\keyword{delete}
