% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{Filter Data Using Lazy Evaluation}
\usage{
filterData(data, filterCat1, value1)
}
\arguments{
\item{data}{This is the data source to be filtered}

\item{filterCat1}{This is the column name which should be filtered}

\item{value1}{This is the value or value list to filter (evaluate to true)}
}
\description{
This function filters a data frame using lazy evaluation.
}
\keyword{data}
\keyword{manipulation}
\keyword{modal,}
