% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterNotData.R
\name{filterNotData}
\alias{filterNotData}
\title{Filter Data Not Matching a Given Criteria Using Lazy Evaluation}
\usage{
filterNotData(data, filterCat1, value1)
}
\arguments{
\item{data}{This is the data source to be filtered}

\item{filterCat1}{This is the column name which should be filtered}

\item{value1}{This is the value or value list to filter (evaluate to false)}
}
\description{
This function filters a data frame using lazy evaluation.
}
\examples{
dat <- filterData(mtcars,"am",c(1,0))
}
\keyword{data}
\keyword{manipulation}
