% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNextID.R
\name{GetNextId}
\alias{GetNextId}
\title{Get Next ID (CRUD)}
\usage{
GetNextId()
}
\description{
This function gets the next ID in a data set as part of a CRUD series of functions.  Used in the create data function.
}
\examples{
CreateData <- function(data) {
data <- CastData(data)
rownames(data) <- GetNextId()
if (exists("responses")) {
  responses <<- rbind(responses, data)
} else {
  responses <<- data
}
}
}
\keyword{CRUD}
