% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableMetadata.R
\name{GetTableMetadata}
\alias{GetTableMetadata}
\title{Get Table Metadata Function (CRUD)}
\usage{
GetTableMetadata()
}
\description{
This function gets table metadata as part of a CRUD series of functions.
}
\examples{
# input fields are treated as a group
formData <- reactive({
  sapply(names(GetTableMetadata()$fields), function(x) input[[x]])
})
}
\keyword{CRUD,}
\keyword{metadata}
