% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharterOptions.R
\name{highcharterOptions}
\alias{highcharterOptions}
\title{Highcharter Options}
\usage{
highcharterOptions()
}
\description{
Sets highcharter options
}
\examples{
highcharterOptions()
}
\keyword{highcharter,}
\keyword{options}
