% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Data from SQL Connection}
\usage{
loadData(sqlString)
}
\arguments{
\item{sqlString}{This is the query that should be sent to the SQL database.}
}
\description{
This function returns a data frame that is the output of a SQL Query.
}
\examples{
  dat <- loadData("SELECT * FROM t_customers")
}
\keyword{RMySQL}
\keyword{sql,}
