% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPkgs.R
\name{loadPkgs}
\alias{loadPkgs}
\title{Load Packages}
\usage{
loadPkgs(pkgList, libPath)
}
\arguments{
\item{pkgList}{Package list to load}

\item{lib_path}{Library path from which to load packages}
}
\description{
This function loads a list of packages from a specified location.
}
\examples{
other.pkgs1 <- c("shiny","tidyr","dplyr","shinydashboard.uber","highcharter",
                 "ggplot2","purrr","scales","tidytext","stringr")
loadPkgs(other.pkgs1, lib_path)
print("All other packages loaded")
}
\keyword{handling}
\keyword{package}
