% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problemPackages.R
\name{problemPackages}
\alias{problemPackages}
\title{Remove problematic functions}
\usage{
problemPackages()
}
\description{
This function removes and then loads known problematic packages from a shiny server instance.
}
\examples{
# input fields are treated as a group
problemPackages()
}
\keyword{handling}
\keyword{package}
