% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{ReadData}
\alias{ReadData}
\title{Read Data Function (CRUD)}
\usage{
ReadData()
}
\description{
This function reads data from a saved source.
}
\examples{
observeEvent(input$responses_rows_selected, {
  if (length(input$responses_rows_selected) > 0) {
    data <- ReadData()[input$responses_rows_selected, ]
    UpdateInputs(data, session)
  }
})
}
\keyword{CRUD,}
\keyword{create}
