% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{Append Data from Data Frame to MySQL Table}
\usage{
saveData(data, table)
}
\arguments{
\item{data}{This is the data frame which has the data to be appended.}

\item{table}{This is the name of the table in MySQL to which you are appending data.}
}
\description{
This function appends data from a data frame to a table in a MySQL database.
}
\examples{
  dat <- formData()
  saveData(dat,"t_customers")
}
\keyword{RMySQL}
\keyword{sql,}
