% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondToNumeric.R
\name{second_el_to_numeric}
\alias{second_el_to_numeric}
\title{Second to Numeric}
\usage{
second_el_to_numeric(ls)
}
\arguments{
\item{ls}{List to convert to numeric.}
}
\description{
This function converts a list to numeric for use with highcharter drilldown charts.
}
\examples{
for (i in 1:length(monthNames)){
  x <- lineData() \%>\%
    filter(MonthName == monthNames[i]) \%>\%
    group_by(AccountDescription) \%>\%
    summarise(value = round(sum(Actual) - sum(Budget),2)) \%>\%
    mutate(name = AccountDescription) \%>\%
    arrange(-value) \%>\%
    select(name, value) \%>\%
    list_parse2() \%>\%
    second_el_to_numeric()

  assign(paste0("ds",tolower(monthNames[i])),x)

  newList <- list(
    id = tolower(monthNames[i]),
    data = get(paste0("ds",tolower(monthNames[i]))),
    name = "Amount"
  )

  assign(paste0("list",i),newList)
}
}
\keyword{highcharter,}
\keyword{numeric}
