% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSQLOptions.R
\name{setSQLOptions}
\alias{setSQLOptions}
\title{Set SQL Options}
\usage{
setSQLOptions(localhoststr, portno, userstr, passwdstr, dbNameStr)
}
\arguments{
\item{dbNameStr}{Database name to use for connections}

\item{host}{Host location (often times localhost)}

\item{port}{Port where the mysql server is open}

\item{user}{MySQL username}

\item{password}{Password used for login for MySQL server}
}
\description{
This function sets SQL default options for making database connections
}
\examples{
  setSQLOptions("localhost",3306,"user","password","dbName")
}
\keyword{RMySQL,}
\keyword{mysql}
\keyword{sql,}
