% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyBoots.R
\docType{package}
\name{shinyBoots-package}
\alias{shinyBoots}
\alias{shinyBoots-package}
\title{shinyBoots Package Documentation}
\description{
This package contains functions to make it easier to create a reactive website
using Twitter Bootstrap and Shiny.  It also contains some functions to make it
easier to connect to local database servers.
}
\details{
The package contains functions that fall into the following general categories:
functions to create boostrap elements, functions to create a reactive app with CRUD
functions (create, retrieve, update, and delete), functions to create charts and
visualizations, and functions to handle CSV files.

\subsection{Bootstrap}{

\code{\link{actionLink}} - creates an action link that can be used to trigger
server-side functions

\code{\link{box}} and \code{\link{boxUI}} - creates box containers in the UI

\code{\link{buttonModal}} - creates a button that opens a modal

\code{\link{buttonRow}} - creates a row of buttons

\code{\link{createLaunchModal}} - creates a modal that launches on web site load

\code{\link{createLoadModalScript}} - creates a modal script to launch a modal on web site load

\code{\link{createPrezModalScript}} - creates a modal that can contain a Google Slides presentation

\code{\link{createUIModal}} - creates a modal that can contain a Shiny UI
}
}
\author{
\strong{Maintainer}: Jordan DeHerrera \email{jordan.deherrera@gmail.com}

}
