% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackingScript.R
\name{trackingScriptOut}
\alias{trackingScriptOut}
\title{Google Analytics Tracking Script}
\usage{
trackingScriptOut(trackingID)
}
\arguments{
\item{trackingID}{This is the tracking ID created by the Google Analytics properties page.}
}
\description{
This function creates a tracking script for Google Analytics.
}
\examples{
tags$head(trackingScriptOut(trackingID))
}
\keyword{tracking}
