% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateData.R
\name{UpdateData}
\alias{UpdateData}
\title{Update Data Function (CRUD)}
\usage{
UpdateData(data)
}
\arguments{
\item{data}{Data to be updated}
}
\description{
This function updates data in a stored source.
}
\examples{
observeEvent(input$submit, {
  if (input$id != "0") {
    UpdateData(formData())
  } else {
    CreateData(formData())
    UpdateInputs(CreateDefaultRecord(), session)
  }
}, priority = 1)
}
\keyword{CRUD,}
\keyword{update}
