% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInputs.R
\name{UpdateInputs}
\alias{UpdateInputs}
\title{Update Inputs Function (CRUD)}
\usage{
UpdateInputs(data, session)
}
\arguments{
\item{data}{Data to be cast.}
}
\description{
This function updates inputs as part of a CRUD series of functions.
}
\examples{
# Click "Submit" button -> save data
observeEvent(input$submit, {
  if (input$id != "0") {
    UpdateData(formData())
  } else {
    CreateData(formData())
    UpdateInputs(CreateDefaultRecord(), session)
  }
}, priority = 1)
}
\keyword{CRUD,}
\keyword{update}
