% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueBoxGroup.R
\name{valueBoxGroup}
\alias{valueBoxGroup}
\title{Value Box Group}
\usage{
valueBoxGroup(..., title = "", width = 6)
}
\arguments{
\item{...}{Content of Value Box Group}

\item{title}{Title of the box}

\item{width}{Width of value box group}
}
\description{
This function creates a value box group.
}
\examples{
valueBoxGroup(boxUI(title = "Test Title", uiOutput("Some Output")))
}
\keyword{UI,}
\keyword{box,}
\keyword{group}
