% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideToLong.R
\name{wideToLong}
\alias{wideToLong}
\title{Convert Dataframe from Wide to Long}
\usage{
wideToLong(data, colSelections)
}
\arguments{
\item{data}{This is the data source to be converted}

\item{colSelections}{These are the columns (references by column number) that contain the values}
}
\description{
This function converts a data frame from wide format to long.
}
\keyword{data}
\keyword{manipulation}
