% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeToSeries.R
\name{changeToSeries}
\alias{changeToSeries}
\title{Convert a Data Frame to a Series for Use in Highcharter}
\usage{
changeToSeries(dat, nameLabel, valueLabel)
}
\arguments{
\item{dat}{This is the data source to be converted}

\item{nameLabel}{This is the name of the column that should be the "name" part of the series}

\item{valueLabel}{This is the name of the column that contains the values that will be the "data" in the series}
}
\description{
This function converts a data frame to a series for use in highcharter
}
\examples{
library(dplyr)
mtcars <- mtcars \%>\% mutate(vehicle_name = row.names(.))
datSeries <- changeToSeries(mtcars,"vehicle_name","mpg")
}
\keyword{data}
\keyword{highcharter,}
\keyword{manipulation,}
\keyword{series}
