% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeToSeries.R
\name{changeToSeries}
\alias{changeToSeries}
\title{Convert a Data Frame to a Series for Use in Highcharter}
\usage{
changeToSeries(df, keyColumnNameAsText)
}
\arguments{
\item{df}{This is the data source to be converted}

\item{keyColumnNameAsText}{This is the name of the column that should be the "name" part of the series}
}
\description{
This function converts a data frame to a series for use in highcharter
}
\examples{
library(dplyr)
dat2 <- prepSeries(mtcars,"cyl","am","mpg") \%>\% mutate(am = ifelse(am == 1,"Automatic","Manual"))
datSeries <- changeToSeries(dat2,"am")
createColumn(datSeries,unique(dat2$cyl),"Fuel Efficiency","Miles per Gallon by Transmission and Cylinders")
}
\keyword{data}
\keyword{manipulation}
