% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyBoots.R
\docType{package}
\name{shinyBoots-package}
\alias{shinyBoots}
\alias{shinyBoots-package}
\title{Generate R documentation from inline comments.}
\description{
This package contains functions to make it easier to create a reactive website
using Twitter Bootstrap and Shiny.  It also contains some functions to make it
easier to connect to local database servers.
}
\details{
Some useful functions in shinyBoots include the following:
\code{\link{actionLink}} - to create a link that will start an action on the server-side
\code{\link{buttonModal}} - to create a button that will open a modal
}
\author{
\strong{Maintainer}: Jordan DeHerrera \email{jordan.deherrera@gmail.com}

}
